[Map Server] Loading of locations, music and Pokecenters lists and sending them to new clients if they connect.

+++ Go/src/pulogic/models/pokecenter.go
@@ -8,6 +8,6 @@ const (
 
 type Pokecenter struct {
 	Idpokecenter int `PK`
-	Position     int
+	Position     int64
 	Description  string
 }
